Add files via upload
I typed this code via C++ using CodeBlocks being that is my most familiar IDE. I had fun creating this program, because I saw it as a swarm of robots waiting for their task. My code is set to run by constantly reading from an inputfile and then creating an outputfile that can replace the previous input file. This input file is set to read from the HTTPS website we were given, but I believe some problems were occurring with this for me sadly. I have not previous interfaced with HTTPS, so if that eliminates me from contention I understand. All I ask is can someone explain how I could made that menu function work. That was the only menu function I could not get to load properly. 

But Regardless thankyou for the challenge! It was fun 
- Somjai Murrell Jr.